Add a test for group independence in capacity checks
In this example, there are two node groups. The second is
not redundant, as it contains only one non-drained node and
an instance (so, should that node fail as well, the instance
cannot stay in the group). However, there are no current
violations; every instance is running on a healthy node.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>