Do not search shrinking most pressing resource twice
The search in tiered allocation is to first try shrinking the instance
to be allocated on the resource that blocks most allocations, and if
that fails try if unilaterally shrinking a single resource leads to a
valid solution. In that second step, we can drop the resource
investigated in the first step, as, after a shrinking step,
unilaterally shrinking it will be considered anyway.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>