Use monotoneFind to improve speed of doShrink
The doShrink function, which is used for tieredAllocation,
tries to find the first shrinking size where the allocation
statistics changes. As computing the allocation statistics
is an expensive operation, use a guess and verify strategy
to do this computation less often.

While there, also fix a typo in the function description.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>