Introduce new failure mode: instance too small
Most failure modes in allocating an instance indicate a lack
of some resource, and we want to know which resource we're running
short of. In this case, shrinking the instance to be allocated
on this resource might solve the problem.

However, there is one case that is qualitatively different: an
allocation might not be possible if the instance is smaller than
the minimal requirement set by the policy. Create a new failure
mode for this case. Besides beeing more informative, we also
ensure that all components of the failure statistics are monotone
(either increasing or decreasing) with the instance size.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>