Also ship the hmac key with the inital data
On setting up a node, initial data is shipped via ssh to the
target node, so that the node daemon can run properly to take
on the remaining configuration tasks via RPC. At that point, also
other daemons, including the monitoring daemon (if enabled) are
started. Therefore, also ship the hmac key file as part of the
initial data, as the monitoring daemon needs it for confd queries.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>