Add a function estimating the redundancy of a group
The redundancy level of a group is the number of nodes that can sequentially
fail while still keeping it possible to have all instances on the cluster.
As per our design, we estimate this level by assuming for an N+1 redundant
cluster that the worst next failure is that of the biggest node.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>