kvm: Refactor device option handling
Replace pci slot with hvinfo in Disk and NIC config Objects. This
will contain all the necessary info to construct a -device option.
Specifically, for PCI devices it will have driver, id, bus, and addr
fields. SCSI devices will have driver, id, bus, channel, scsi-id,
and lun fields.

Change the way we generate the device ID so that it gets derived
only from the type and the UUID of the device (e.g.
disk-932df160-7a22-4067). We ensure seamless migration by upgrading
the existing runtime files in _UpgradeSerializedRuntime(). All
devices found with a 'pci' slot will get an 'hvinfo' dict with the
old id, bus 'pci.0', and addr based on the old value of 'pci'.

Change the way we put devices into buses. Guessing the default
reserved PCI slots of an instance has proven to be error-prone,
since they may change from version to version, and also because the
PCI bus state may be modified with a specific option passed by the
user with the kvm_extra hvpamam. Therefore, we decide to start
adding PCI devices from slot 12 onwards. The SCSI bus does not need
to have default reservations.

Let QEMU decide the PCI slot of spice, balloon and SCSI controller
devices. Finally, QEMU adds by default an LSI controller in case a
SCSI disk is attached to the instance. Here, we set it explicitly,
and name the created bus 'scsi.0' (as QEMU does by default).

Until now, only SCSI emulated disks were supported. Allow
scsi-generic, scsi-block, scsi-hd, and scsi-cd for the disk_type
hvparams. The first two do SCSI pass-through and thus require a
physical SCSI device on the host. The third is the equivalent of
if=scsi, which means that the virtual disk of the instance will be an
emulated SCSI device.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>