monitor: Use hvinfo in QMP methods
Change hotplug-related methods (i.e., HotAddNic() and HotAddDisk())
to use slot 'hvinfo' instead of slot 'pci'.

Change HasPCIDevice() to HasDevice(). This will parse the output of
the 'query-block' and 'query-pci' QMP commands and try to match the
given device ID.

Make _VerifyHotplugCommand() depend only on the device ID and not
on the whole device object.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>