kvm: Work around QEMU commit 48f364dd
QEMU commits 48f364dd and da2cf4e8 included in version 2.2 and later
state the following:

 * HMP's drive_del is not supported any more on a drive added
   via QMP's blockdev-add

 * Stay away from immature blockdev-add unless you want to help
   with development.

To this end, currently we cannot hot-remove a disk that was
previously hot-added. Here, we work around this constrain by using
HMP's drive_add instead of blockdev-add. This must be done via a
callback in HotAddDisk() wrapper method, due to the fact that if a
QMP connection terminates before a drive keeps a reference to the fd
passed via the add-fd QMP command, then the fd gets closed and
cannot be used later. So we open a QMP connection, pass the fd, then
invoke the drive_add HMP command referencing the corresponding
fdset, then remove the fdset, then add the device and then terminate
the QMP connection.

Signed-off-by: Dimitris Aragiorgis <dimitris.aragiorgis@gmail.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>