Add a utility function for the path to kvm's pid files
The process IDs for the kvm processes controlled by Ganeti
are stored in pid files, explicitly specified for Ganeti.
While the canonical reference for those files is in the kvm
hypervisor code (lib/hypvervisor/hv_kvm/__init__.py), also
provide that path in the Haskell world via the path utils.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>