Add a new option for assuming idleness
When dynamic utilization information for a resource is not
available, the default value is used, which is maximal utilization.
This effectively makes hbal spread out an equal number of instances
per node. Of course, dynamic utilization can be ignored all together
by the --ignore-dynu option. However, neither covers the use case
only one resource is scarce and hence should be balanced optimally,
whereas the other resources are not measured, but still should be
ignored. Therefore, add a new option telling hbal to assume 0 utilization
for all resources where no data is provided.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Oleg Ponomarev <oponomarev@google.com>