Support scaling of the memory weight
While the weights of the metric components cannot be changed easily
at run time, the specification for dynamic utilization is that they be
in arbitrary units. Therefore, simply scaling the utilization values
changes the weight of this measure. Note that it still makes sense to
have as a default measure the percentage of memory used.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Oleg Ponomarev <oponomarev@google.com>