Increase the weight of the memory balance
Already simple, examples (like, e.g., the recently added unit tests)
show that a weight of 1.0 for the memory balance is not high enough
to influence a decision against static balancing aspects; this is also
to be expected as there are several of them (static memory, static disk,
static CPU, etc) of total weight 1.0 each. Therefore, increase the weight
that the maintenance daemons assigns to this component on memory over
committed clusters significantly and make it increase with increasing
over commitment. The function is still guesswork, but should be a much
better starting point.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Oleg Ponomarev <oponomarev@google.com>