Reduce flakyness of GetCmdline test on slow machines
The GetCmdline test verifies that we can get the command line
of a running process via the procfs. To not have to care about
cleanup, the test creates an ephemeral process for this. While
two wall-clock seconds seem more than enough for a single read
from the procfs on nowadays machines, this is not true for some
of the public buildbot (virtual) machines which are extremely
low on resources and can have really heavy load; this causes
flakyness of that test there. Mitigate this by increasing the
life time of the process.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>