Extend timeout for gnt-cluster renew-crypto
With particularly large clusters, the renewal of SSH keys happening in
renew-crypto can take a long time to complete. While this should be
improved, an additional problem is that the RPC doing most of the work
has a default one-hour timeout. Given that it is preferable that the
operation completes, this patch bumps the timeout to four hours, which
should suffice even for 80+ node clusters.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Oleg Ponomarev <oponomarev@google.com>