Disable --no-node-setup
This patch disables the option --no-node-setup of
'gnt-node add'. The option is the equivalent to
--no-ssh-init for 'gnt-cluster init'. However, it
was rather cumbersome for users to remember whether
or not the cluster was initialized with that to
not forget to use this option. Instead making the
user use --no-node-setup, Ganeti shall determine
whether or not to change the SSH setup by reading
the configuration parameter 'modify_ssh_setup' which
is set with --no-ssh-init.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>