Always pass node uuids but not node names to hooksmaster
Currently hooksmaster is able to deal with both uuids and names.
While uuids are passed in most cases, the node list uses the legacy
naming *node_names* inside the hooksmaster implementation which leads
to misunderstandings. This patch modifies the existing code to make
it always pass node uuids to hooksmaster. It also renames the
misleading *node_names* parameter to *node_uuids*.

Signed-off-by: Oleg Ponomarev <oponomarev@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>