Inform users of timed out cancelations
When multiple job cancel operations are executed at once, there is a
chance for the contention to result in a job still in the canceling
state at timeout time. Ganeti would emit a funny message:

"Job could not be canceled, status JOB_STATUS_CANCELING"

This patch explains what is happening, making the message a bit more
user-friendly. It also makes all the related output consistently use
the American spelling of canceled.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>