Add an authenticated user name to an opcode's reason trail
The entry will have the following format:
("gnt:library:rlib2", "RAPI-Auth:user", *time*).

This entry in the reason trail might be useful in order to detect
malicious rapi users or restrict access for some operations by filters.
Note that, in the last case, it's necessary to restrict access to
filters modification first.

Signed-off-by: Oleg Ponomarev <oponomarev@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>