Add storage ids to basic RAPI GetInstance response
We add a new field storage_ids to the RAPI output of basic instance
information. This information is retrieved from the instance disks'
logical ids, so that for each disk it is either a part of the storage
path or a unique name used for external storage. Note that there will be
no value returned for DRBD disks.

Signed-off-by: Lisa Velden <velden@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>