iallocator: only adjust memory usage for up instances
With the introduction of forthcoming instances, htools will
consider two memory states: the memory usage now, and the
future one, when all forthcoming one, when all instances are
created and started. In particular, it is not necessary,
and in fact harmful, to reserve the memory of currently down
instances by subtracting it from the current free memory at
the IAllocator interface. IAllocators are supposed to take future
memory usage into account in the same way as they are supposed
to ensure N+1 redundancy.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>

Cherry-picked-from: eacfdba57f24
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Brian Foley <bpfoley@google.com>